test: file the MultipartScanSession suite under its own subject
It sat at the bottom of a 1,528-line camera-lifecycle file, which is where the
relay parser tests sat before they were moved for the same reason: filed by
proximity rather than by what they assert.
